pyproject.toml
[project]
name = "fault-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "pandas>=2.2.3",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
